renames package to download_directory
Underscore to replace dash: https://stackoverflow.com/a/3960773/2295778
