+++ src/main/java/org/reactome/release/download_directory/Biopax.java
@@ -1,4 +1,4 @@
-package org.reactome.release.downloadDirectory;
+package org.reactome.release.download_directory;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.logging.log4j.LogManager;
