+++ src/main/java/org/reactome/release/download_directory/CreateReactome2BioSystems.java
@@ -1,4 +1,4 @@
-package org.reactome.release.downloadDirectory;
+package org.reactome.release.download_directory;
 
 import java.io.File;
 import java.io.IOException;
@@ -21,7 +21,7 @@ public class CreateReactome2BioSystems {
 		final Path bioSystemsTempFilePath = Paths.get(BIO_SYSTEMS_TEMP_DIRECTORY, bioSystemsFile);
 		final Path bioSystemsTargetFilePath = Paths.get(releaseNumber, bioSystemsFile);
 		logger.info("Running CreateReactome2BioSystems step");
-		// The last argument, 'BioSystems', specifies the output directory of the ReactomeToBioSystems.zip file. 
+		// The last argument, 'BioSystems', specifies the output directory of the ReactomeToBioSystems.zip file.
 		// The script removes all files within the named directory, so only change the output directory to an empty or non-existent one
 		ReactomeToBioSystemsConverter.main(new String[] {host, database, username, password, Integer.toString(port), BIO_SYSTEMS_TEMP_DIRECTORY});
 		Files.move(bioSystemsTempFilePath, bioSystemsTargetFilePath, StandardCopyOption.REPLACE_EXISTING);
