+++ src/main/java/org/reactome/release/download_directory/GenerateGOAnnotationFile/CellularComponentAnnotationBuilder.java
@@ -1,11 +1,11 @@
-package org.reactome.release.downloadDirectory.GenerateGOAnnotationFile;
+package org.reactome.release.download_directory.GenerateGOAnnotationFile;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.gk.model.GKInstance;
 import org.gk.model.ReactomeJavaConstants;
 
-import static org.reactome.release.downloadDirectory.GenerateGOAnnotationFile.GOAGeneratorConstants.*;
+import static org.reactome.release.download_directory.GenerateGOAnnotationFile.GOAGeneratorConstants.*;
 
 import java.util.*;
