+++ src/main/java/org/reactome/release/download_directory/GenerateGOAnnotationFile/MolecularFunctionAnnotationBuilder.java
@@ -1,4 +1,4 @@
-package org.reactome.release.downloadDirectory.GenerateGOAnnotationFile;
+package org.reactome.release.download_directory.GenerateGOAnnotationFile;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -6,7 +6,7 @@ import org.gk.model.GKInstance;
 import org.gk.model.ReactomeJavaConstants;
 import org.gk.schema.SchemaClass;
 
-import static org.reactome.release.downloadDirectory.GenerateGOAnnotationFile.GOAGeneratorConstants.*;
+import static org.reactome.release.download_directory.GenerateGOAnnotationFile.GOAGeneratorConstants.*;
 
 import java.util.*;
