+++ src/main/java/org/reactome/release/download_directory/GenerateGOAnnotationFile/README.md
