+++ src/main/java/org/reactome/release/download_directory/HumanPathwaysWithDiagrams.java
@@ -1,4 +1,4 @@
-package org.reactome.release.downloadDirectory;
+package org.reactome.release.download_directory;
 
 import org.gk.model.GKInstance;
 import org.gk.model.ReactomeJavaConstants;
