+++ src/test/java/org/reactome/release/download_directory/GenerateGOAnnotationFile/MolecularFunctionAnnotationBuilderTest.java
@@ -1,4 +1,4 @@
-package org.reactome.release.downloadDirectory.GenerateGOAnnotationFile;
+package org.reactome.release.download_directory.GenerateGOAnnotationFile;
 
 import org.gk.model.GKInstance;
 import org.gk.model.ReactomeJavaConstants;
